Config values fail type assertion to uint, so treat them as int and then convert to uint when creating worker.Config objects

+++ cmd/token-push/main.go
@@ -297,7 +297,7 @@ func run(ctx context.Context) error {
 			return errors.New(msg)
 		}
 		workerRetryMap[retryArg.WorkerType] = workerRetryConfig{
-			numRetries: numRetries,
+			numRetries: uint(numRetries),
 			retrySleep: retrySleep,
 		}
 	}
